==> R/5-dashboard.Rmd <==
---
title: "PigPeaks Dashboard"
output:
  flexdashboard::flex_dashboard:
    theme: simplex
    vertical_layout: scroll
    social: menu
    orientation: rows
    df_print: paged
#    logo: ../SVA_logo.png
---

<style type="text/css">

.html-widget.gauge svg {
  height:20%; /*or try sth like 320px instead of 100%, whatever you prefer*/
  margin-top: -1px;
  margin-bottom: -1px;
}

.chart-title {  /* chart_title  */
   font-size: 17px;
}

.navbar {
  font-size:17px;

}


.tabset { 
  font-size:17px;
}

.dropdown-menu {
  font-size:15px;
}

</style>

```{r message=FALSE, warning=FALSE}
library(flexdashboard)
library(shiny)
library(dplyr)
library(plotly)
library(ggplot2)
library(stringr)


setwd('..') #do not run this line if running manually (instead of knitting)


source('Settings.r')
source('R/Functions.r') 
source('R/Functions-plotting.r')
load('data/indicators.RData')
load('data/indicators.results.RData')




```


```{r}

# function that chose the correct arguments for each indicator plot
chose.arguments.plot <- function(i = index
){
  
indicator.label = "Observed"

if (isTRUE(indicators.sys[i])) {
  
  vertical.line = NULL
  vertical.line.label=NULL
  shading.matrix = NULL
  limits = NULL
  series.line = NULL
  group.labels=c('gilts','young','prime','mature')
  y2label = '% of'
  y2range = c(0,1)
  
  if (indicators.limits[i]=="limit.upp") {
    
    UCL.EWMA = TRUE
    LCL.EWMA = FALSE
    UCL.SHEW = TRUE
    LCL.SHEW = FALSE
    alarms.EWMA.UPP = TRUE
    alarms.EWMA.LW = FALSE
    alarms.SHEW.UPP = TRUE
    alarms.SHEW.LW = FALSE
    
  }
  if (indicators.limits[i]=="both") {
    
    UCL.EWMA = TRUE
    LCL.EWMA = TRUE
    UCL.SHEW = FALSE
    LCL.SHEW = FALSE
    alarms.EWMA.UPP = TRUE
    alarms.EWMA.LW = TRUE
    alarms.SHEW.UPP = TRUE
    alarms.SHEW.LW = TRUE
  }
  
  if (indicators.limits[i]=="limit.lw") {
    
    UCL.EWMA = FALSE
    LCL.EWMA = TRUE
    UCL.SHEW = FALSE
    LCL.SHEW = TRUE
    alarms.EWMA.UPP = FALSE
    alarms.EWMA.LW = TRUE
    alarms.SHEW.UPP = FALSE
    alarms.SHEW.LW = TRUE
  }
  
}
if (isFALSE(indicators.sys[i])) {
  
  vertical.line = NULL
  vertical.line.label = NULL
  shading.matrix = NULL
  limits = NULL
  series.line = NULL
  group.labels=c('gilts','young','prime','mature')
  y2label = '% of'
  y2range = c(0,1)
  UCL.EWMA = FALSE
  LCL.EWMA = FALSE
  UCL.SHEW = FALSE
  LCL.SHEW = FALSE
  alarms.EWMA.UPP = FALSE
  alarms.EWMA.LW = FALSE
  alarms.SHEW.UPP = FALSE
  alarms.SHEW.LW = FALSE
  
}

if (indicators.type[i]=="W") {
  
  show.window = weeks.to.show
  index.dates = index.dates.week
  xlabel = 'Week'
  
}

if (indicators.type[i]=="C") {
  
  show.window = nonTS.to.show
  index.dates = index.dates.days
  xlabel = 'Events'
  
}

arguments <- list(indicator.label=indicator.label, vertical.line=vertical.line,
                  vertical.line.label=vertical.line.label, shading.matrix=shading.matrix,
                  limits=limits, series.line=series.line, group.labels=group.labels, UCL.EWMA=UCL.EWMA,
                  LCL.EWMA=LCL.EWMA, UCL.SHEW=UCL.SHEW, LCL.SHEW=LCL.SHEW,
                  alarms.EWMA.UPP=alarms.EWMA.UPP, alarms.EWMA.LW=alarms.EWMA.LW,
                  alarms.SHEW.UPP=alarms.SHEW.UPP, alarms.SHEW.LW=alarms.SHEW.LW,
                  show.window=show.window, index.dates=index.dates, xlabel=xlabel, 
                  y2label=y2label, y2range=y2range)


return(arguments)
}
  
```

# Overview {.sidebar data-width=200}


Management {data-navmenu="Gilts"}
====================================

```{r, results='asis'}

if("time.to.first.service" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="time.to.first.service")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")

  nonTS.barplot.timeless(df.indicator = indicators.time.series[[i]],
                         argument.list = TRUE,
                         target = NULL,
                         target.unit= NULL, 
                         ylabel = "time (in days)",
                         series.label = "gilts")
}
```


```{r, results='asis'}

if("time.to.first.farrowing" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="time.to.first.farrowing")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.timeless(df.indicator = indicators.time.series[[i]],
                         argument.list = TRUE,
                         target = NULL,
                         target.unit= NULL,
                         ylabel = "time (in days)",
                         series.label = "gilts")
}

```


Health {data-navmenu="Empty Sows"}
====================================

```{r, results='asis'}

if("empty.long.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="empty.long.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "sows",
                use.minimum.y="zero")
}
```


Health {data-navmenu="Services"}
====================================

```{r, results='asis'}

if("reservices.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="reservices.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "reservices",
                use.minimum.y="zero")
}

```


```{r, results='asis'}

if("time.to.reservice" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="time.to.reservice")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "time (in days)",
                            use.minimum.y = "min")
  
}

```

```{r, results='asis'}

if("time.to.reservice" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="time.to.reservice")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("time.to.reservice" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="time.to.reservice")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("time.to.reservice" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="time.to.reservice")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average time (in days)",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("rereservices" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="rereservices")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "nº of sows", 
                            use.minimum.y = "min")
}

```

```{r, results='asis'}

if("rereservices" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="rereservices")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("rereservices" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="rereservices")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("rereservices" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="rereservices")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of sows",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.failure" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.failure")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
    TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                  argument.list = TRUE,
                  target = NULL,
                  target.unit = NULL,
                  ylabel = "% failure",
                  use.minimum.y="zero")
}

```


```{r, results='asis'}

if("perc.reservice" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="perc.reservice")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% reservice",
                use.minimum.y="zero")
}

```


Management {data-navmenu="Services"}
====================================

```{r, results='asis'}

if("services.week" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="services.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "services",
                use.minimum.y="min")
}
```


Health {data-navmenu="Pregnancy"}
====================================

```{r, results='asis'}

if("perc.pregnancy" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.pregnancy")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% pregnancy",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("abortions.week" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="abortions.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "abortions",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("services.to.farrow" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="services.to.farrow")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "nº of services to farrow",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("services.to.farrow" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="services.to.farrow")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("services.to.farrow" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="services.to.farrow")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("services.to.farrow" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="services.to.farrow")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of services to farrow",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("pregnancy.length" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="pregnancy.length")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,  
                            ylabel = "pregnancy length (in days)",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("pregnancy.length" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="pregnancy.length")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("pregnancy.length" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="pregnancy.length")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("pregnancy.length" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="pregnancy.length")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average pregnancy length (in days)",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


Management {data-navmenu="Pregnancy"}
====================================

```{r, results='asis'}

if("time.to.abortion" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="time.to.abortion")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "time (in days)",
                            use.minimum.y = "min")
}
```


Health {data-navmenu="Farrowing"}
====================================

```{r, results='asis'}

if("days.between.farrowings" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="days.between.farrowings")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "nº of days",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("days.between.farrowings" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="days.between.farrowings")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("days.between.farrowings" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="days.between.farrowings")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("days.between.farrowings" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="days.between.farrowings")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of days",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("farrowings.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="farrowings.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of farrowings",
                use.minimum.y="zero")
}
```


```{r, results='asis'}

if("total.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}

```


```{r, results='asis'}

if("live.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="live.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("dead.born.week" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="dead.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("small.born.week" %in% names(indicators.data)) {

  i = which(names(indicators.data)=="small.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("weak.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weak.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("mummi.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="mummi.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "nº of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("perc.dead.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.dead.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("perc.small.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.small.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("perc.weak.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.weak.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("perc.mummi.born.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.mummi.born.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "% of piglets",
                use.minimum.y="min")
}
```


```{r, results='asis'}

if("total.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("total.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("total.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("total.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("live.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="live.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("live.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="live.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("live.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="live.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("live.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="live.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="dead.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,  
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="dead.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="dead.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="dead.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.dead.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "% of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("perc.dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.dead.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("perc.dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.dead.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("perc.dead.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.dead.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average % of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="small.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,  
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="small.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="small.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="small.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.small.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "% of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("perc.small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.small.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("perc.small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.small.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("perc.small.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.small.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average % of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weak.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weak.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weak.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weak.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.weak.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "% of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("perc.weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.weak.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("perc.weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.weak.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("perc.weak.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.weak.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average % of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="mummi.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="mummi.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="mummi.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="mummi.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.mummi.born.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,   
                            ylabel = "% of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("perc.mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.mummi.born.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("perc.mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.mummi.born.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("perc.mummi.born.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.mummi.born.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average % of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


Health {data-navmenu="Weaning"}
====================================

```{r, results='asis'}

if("total.wean.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.wean.week")

  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "piglets",
                use.minimum.y="zero")
}
```


```{r, results='asis'}

if("diff.wean.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="diff.wean.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "piglets",
                use.minimum.y="zero")
}
```


```{r, results='asis'}

if("negdiff.wean.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="negdiff.wean.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "piglets",
                use.minimum.y="zero")
}
```


```{r, results='asis'}

if("total.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.wean.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "nº of piglets",                    
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("total.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.wean.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("total.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.wean.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("total.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="total.wean.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="negdiff.weaned.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "nº of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="negdiff.weaned.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="negdiff.weaned.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="negdiff.weaned.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average nº of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("perc.negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.negdiff.weaned.litter")

  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
    nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL, 
                            ylabel = "% of piglets",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("perc.negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.negdiff.weaned.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("perc.negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.negdiff.weaned.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("perc.negdiff.weaned.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="perc.negdiff.weaned.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "average % of piglets",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


```{r, results='asis'}

if("weight.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weight.wean.litter")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,  
                            ylabel = "weight (in kg)",
                            use.minimum.y = "min")
}
```

```{r, results='asis'}

if("weight.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weight.wean.litter")
  
    cat("  \n###", indicators.labels[i],", number of events and alarms per year","  \n")
  
    TS.barplot.continuous.events(df.indicator=indicators.continuous.to.weekly[[i]],
                                 years.to.see = 3,
                                 limits=indicators.limits[[i]])
}
```

```{r, results='asis'}

if("weight.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weight.wean.litter")
  
  cat("  \n## {data-height=450}")
  cat("  \n###", indicators.labels[i],", per week","  \n")
  cat("In the graph on the side, parity is represented as a moving average of",moving_average, "weeks, used to identify trend direction.")
}
  
```

```{r, results='asis'}

if("weight.wean.litter" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weight.wean.litter")
  
  cat("  \n###", indicators.labels[i],", per week","  \n")
  
  TS.barplot.pg.continuous(df.indicator = indicators.continuous.to.weekly[[i]],
                           target = NULL,
                           target.unit = NULL, 
                           ylabel = "weight (in kg)",
                           use.minimum.y = "min",
                           moving_average = moving_average)
}
```


Management {data-navmenu="Weaning"}
====================================

```{r, results='asis'}

if("weanings.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weanings.week")

  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "weanings",
                use.minimum.y="zero")
}
```


```{r, results='asis'}

if("weaning.length" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="weaning.length")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  nonTS.barplot.pg.timeless(df.indicator = indicators.time.series[[i]],
                            argument.list = TRUE,
                            target = NULL,
                            target.unit = NULL,
                            ylabel = "days",                     
                            use.minimum.y = "min")
}
```


Health {data-navmenu="Post-Weaning"}
====================================

```{r, results='asis'}

if("piglets.deaths.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="piglets.deaths.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot(df.indicator = indicators.time.series[[i]],
             argument.list = TRUE,
             target = NULL,
             target.unit = NULL, 
             series.label = "weaned piglets",
             ylabel = "deaths")
}

```


Health {data-navmenu="Exit"}
====================================

```{r, results='asis'}

if("number.deaths.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="number.deaths.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "sows",
                use.minimum.y="zero")
}

```


```{r, results='asis'}

if("gilts.deaths.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="gilts.deaths.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")

  TS.barplot(df.indicator = indicators.time.series[[i]],
             argument.list = TRUE,
             target = NULL,
             target.unit = NULL, 
             series.label = "gilts",
             ylabel = "gilts")
}

```


```{r, results='asis'}

if("number.exits.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="number.exits.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")
  
  TS.barplot.pg(df.indicator = indicators.time.series[[i]],
                argument.list = TRUE,
                target = NULL,
                target.unit = NULL,
                ylabel = "exits",
                use.minimum.y="zero")
}

```


Management {data-navmenu="Exit"}
====================================

```{r, results='asis'}

if("death.after.event.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="death.after.event.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")

  TS.exit(df.indicator=indicators.time.series[[i]],
          indicator.label="Number of sows",
          show.window = weeks.to.show,
          index.dates = index.dates.week,
          ylabel = "death after event",
          xlabel = 'Week',
          target = NULL,
          target.unit = NULL,
          shading.matrix = NULL,
          limits = NULL,
          group.labels = c('birth','service','reservice','abortion','farrowing','weaning'))
}
```


```{r, results='asis'}

if("exit.type.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="exit.type.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")

  TS.exit(df.indicator=indicators.time.series[[i]],
          indicator.label="Number of sows",
          show.window = weeks.to.show,
          index.dates = index.dates.week,
          ylabel = "exit type",
          xlabel = 'Week',
          target = NULL,
          target.unit = NULL,
          shading.matrix = NULL,
          limits = NULL,
          group.labels = c('slaughtered','sold','euthanized','dead','sold.pregnant','missing','exported'))
}
```


```{r, results='asis'}

if("exit.after.event.week" %in% names(indicators.data)) {
  
  i = which(names(indicators.data)=="exit.after.event.week")
  
  cat("  \n##")
  cat("  \n###", indicators.labels[i],"  \n")

  TS.exit(df.indicator=indicators.time.series[[i]],
          indicator.label="Number of sows",
          show.window = weeks.to.show,
          index.dates = index.dates.week,
          ylabel = "exit after event",
          xlabel = 'Week',
          target = NULL,
          target.unit = NULL,
          shading.matrix = NULL,
          limits = NULL,
          group.labels = c('birth','service','reservice','abortion','farrowing','weaning'))
}
```
